reintroducing separate files and query context
