+++ src/SolidQueryDevtools.jsx
@@ -1,8 +1,11 @@
 import Tool from "./Tool";
+import { QueryProvider } from "./Context";
 
 export default function SolidQueryDevtools() {
 
   return (
+    <QueryProvider>
       <Tool />
+    </QueryProvider>  
   );
 };
